pyproject.toml
[tool.poetry]
name = "Akane"
version = "0.2.0"
description = "A general purpose and fun Discord bot."
authors = ["AbstractUmbra <Umbra@AbstractUmbra.xyz>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", rev = "master"}
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus", rev = "master"}
jishaku = {git = "https://github.com/Gorialis/jishaku", rev = "master"}
asyncpg = "*"
googletrans = "3.1.0a0"
psutil = "*"
click = "*"
parsedatetime = "*"
lru_dict = "*"
lxml = "*"
pytz = "*"
python-dateutil = "*"
fuzzywuzzy = "*"
pygit2 = "*"
"mystbin.py" = "*"
feedparser = "*"
CurrencyConverter = "*"
attrdict = "*"
"async-nekos.life-wrapper" = {git = "https://github.com/Nekos-life/Async-nekos.life-wrapper", rev = "master"}
"discord.py-stubs" = "*"
Pillow = "*"
nhentaio = "*"
pykakasi = "*"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
